--- src/_components/headerComponent.js
@@ -1,114 +1,0 @@
-import React from "react";
-import PropTypes from "prop-types";
-import { makeStyles } from "@material-ui/core/styles";
-import AppBar from "@material-ui/core/AppBar";
-import Tabs from "@material-ui/core/Tabs";
-import Tab from "@material-ui/core/Tab";
-import Typography from "@material-ui/core/Typography";
-import Box from "@material-ui/core/Box";
-import LanguageIcon from "@material-ui/icons/Language";//language icon
-import ReactDOM from "react-dom";
-import Dropdown from "muicss/lib/react/dropdown";
-import DropdownItem from "muicss/lib/react/dropdown-item";
-
-function TabPanel(props) {
-    const { children, value, index, ...other } = props;
-
-    return (
-        <Typography
-            component="div"
-            role="tabpanel"
-            hidden={value !== index}
-            id={`nav-tabpanel-${index}`}
-            aria-labelledby={`nav-tab-${index}`}
-            {...other}
-        >
-            {value === index && <Box p={3}>{children}</Box>}
-        </Typography>
-    );
-}
-
-TabPanel.propTypes = {
-    children: PropTypes.node,
-    index: PropTypes.any.isRequired,
-    value: PropTypes.any.isRequired
-};
-
-function a11yProps(index) {
-    return {
-        id: `nav-tab-${index}`,
-        "aria-controls": `nav-tabpanel-${index}`
-    };
-}
-
-function LinkTab(props) {
-    return (
-        <Tab
-            component="a"
-            onClick={event => {
-                event.preventDefault();
-            }}
-            {...props}
-        />
-    );
-}
-
-const useStyles = makeStyles(theme => ({
-    root: {
-        flexGrow: 1,
-        backgroundColor: theme.palette.background.paper
-    }
-}));
-
-export default function NavTabs() {
-    const classes = useStyles();
-    const [value, setValue] = React.useState(0);
-
-    const handleChange = (event, newValue) => {
-        setValue(newValue);
-    };
-
-    return (
-        <div className={classes.root}>
-            <AppBar position="static">
-                <Tabs
-                    variant="fullWidth"
-                    value={value}
-                    onChange={handleChange}
-                    aria-label="nav tabs example"
-                >
-                    <LinkTab label="REVITALIZE" href="/drafts" {...a11yProps(0)} />
-                    <LinkTab label="About Us" href="/drafts" {...a11yProps(1)} />
-                    <LinkTab label="Support" href="/trash" {...a11yProps(2)} />
-                    <LinkTab label="Contact" href="/spam" {...a11yProps(3)} />
-                    <Tab icon={<LanguageIcon />} aria-label="Language">
-                        <Dropdown
-                            color="primary"
-                            icon={<LanguageIcon />}
-                            //label = "Language"
-                            onClick={function () { console.log('toggle clicked') }}
-                            onSelect={function (val) { console.log(val); }} >
-                            <DropdownItem>Language 1</DropdownItem>
-                            <DropdownItem>Language 2</DropdownItem>
-                        </Dropdown>
-                    </Tab>
-                </Tabs>
-            </AppBar>
-            <TabPanel value={value} index={0}>
-                REVITALIZE Home page
-      </TabPanel>
-            <TabPanel value={value} index={1}>
-                About Us page
-      </TabPanel>
-            <TabPanel value={value} index={2}>
-                Customer Support
-      </TabPanel>
-            <TabPanel value={value} index={3}>
-                Contact Us
-      </TabPanel>
-            <TabPanel value={value} index={4}>
-                Language
-      </TabPanel>
-        </div>
-    );
-}
